'changes made ~ overnight'

+++ pages/editrecords_assessment.php
@@ -49,7 +49,7 @@
                   <?php
                     require_once('connection.php');
                     $cn   = new connection();
-                    $conn = $cn->connectDB();
+                    $conn = $cn->connectDB($_SESSION['database']);
 
 
 										$query = 'SELECT * FROM fees';
@@ -72,7 +72,7 @@
                                                 <a href='edit_assessment.php?feecode=<?php echo $row[0]; ?>' class='btn btn-default' style='background-color: #69EC6B;'> Edit </a>
                                             </td>
                                             <td>
-                                                <a href='delete_assessment.php?feecode=<?php echo $row[0]; ?>' class='btn btn-default' style='background-color: #EA6565;'> Delete </a>
+                                                <a href='delete_assessment.php?feecode=<?php echo $row[0]; ?>' class='btn btn-default'  style='background-color: #EA6565;' onclick="return confirm('Are you sure you want to delete this record?');"> Delete </a>
                                             </td>
                                         </tr>
                                         <?php
